requirements.txt
biopython==1.68
click==6.7
Flask==0.11.1
Flask-SQLAlchemy==2.2
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
mock==2.0.0
numpy==1.12.1
pandas==0.20.1
pbr==3.0.1
psycopg2==2.7.1
python-dateutil==2.6.0
pytz==2017.2
six==1.10.0
SQLAlchemy==1.1.10
Werkzeug==0.12.2
